opentapp/requirements.txt
Django==1.6.1
dj-static==0.0.5
dj-database-url==0.2.2
gunicorn==18.0
psycopg2==2.5.1
static==0.4
wsgiref==0.1.2
PyRSS2Gen==1.0.0
South==0.7.6
argparse==1.2.1
bdist-mpkg==0.4.4
colorama==0.2.4
geopy==0.95
py==1.4.12
pyOpenSSL==0.12
pytest==2.3.4
python-dateutil==1.5
python-gcm==0.1.4
pytz==2012h
requests==0.14.2
xattr==0.6.1
zope.interface==3.5.1
PIL==1.1.7
django-widget-tweaks==1.3
djangorestframework==2.3.1
django-simple-captcha==0.4.2
raven==4.0.4
django-extensions